Cleaned up configuration and added examples
All configuration information is now in the env variable.

ldp-app now instantiates and initializes the database and sets it in the env for other services to use. This is to expose database management to the app so it can configure different routes on different database implementations.

Changed the order of the vis and ldp-service routes so ldp-service context can be /, and the root container be a BasicContainer.

vis.js still needs to be implemented on rdflib.